refactor: add env of allowed host and routing filename
